LU-9354 utils: fix lfs.c compile for PPC
Fix the lfs_setstripe_args parameter to use "unsigned long long"
for lsa_comp_end so that when passed to llapi_parse_size() it is
the correct type.  PPC annoyingly uses "unsigned long" for __u64
while most other architectures use "unsigned long long".

Test-Parameters: trivial
Reviewed-on: https://review.whamcloud.com/26727
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>

+++ lustre/utils/lfs.c
@@ -1048,7 +1048,7 @@ static int verify_pool_name(char *prog_name, char *pool_name)
 }
 
 struct lfs_setstripe_args {
-	__u64			 lsa_comp_end;
+	unsigned long long	 lsa_comp_end;
 	unsigned long long	 lsa_stripe_size;
 	int			 lsa_stripe_count;
 	int			 lsa_stripe_off;
